Changed the way in which JOSM handles imagery layer blacklisting. Instead of a hard-coded list of Google URLs, we now parse the server's /api/capabilities response which is expected to look like this:
<osm version="0.6" generator="OpenStreetMap server">
  <api>
    <version minimum="0.6" maximum="0.6"/>
    <area maximum="0.25"/>
    <tracepoints per_page="5000"/>
    <waynodes maximum="2000"/>
    <changesets maximum_elements="50000"/>
    <timeout seconds="300"/>
  </api>
  <policy>
    <imagery>
      <blacklist regex=".*\.google\.com/.*"/>
      <blacklist regex=".*209\.85\.2\d\d.*"/> 
      <blacklist regex=".*209\.85\.1[3-9]\d.*"/>
      <blacklist regex=".*209\.85\.12[89].*"/>
    </imagery>
  </policy>
</osm>

JOSM will now try to establish an API connection when started, so that
it knows about blacklisted layers. It will also re-read the list when 
the URL is changed in the preferences, and if any prohibited layers are
active they will be removed. 

For an interim period, JOSM still uses the four regular expressions
listed above whenever the server API URL is *.openstreetmap.org and
the API does not send any blacklist entries. It is expected that the
API will soon return proper blacklist entries.

Things that could be improved:

1. Establish a general listener system where components can register
their interest in a change of the configured OSM server. Currently 
we have to plug through to the gui layer (Main.main.mapView) from
the base comms layer (OsmApi) which is ugly.

2. Establish a new class of blacklist which works by IP number and not
just regular expression, so that you can say "any host name that resolves
to this IP is blacklisted".

3. Track all layers that were used in editing a data set, and refuse 
uploading if the upload URL bans any of these layers. 



git-svn-id: https://josm.openstreetmap.de/svn/trunk@3934 0c6e7542-c601-0410-84e7-c038aed88b3b