mapcss: change semantics of casing-width: X;
Previously it was an absolute width, from now it's the width of the border on both sides of the main line, so:
width: 6; casing-width: 8 becomes width: 6; casing-width: 1; (or width: 6; casing-width: +2;).

This breaks existing styles! For compatibility with old JOSM versions use the "+" syntax (which is unchanged) or work around with "::layer".
Reason: For most of the other mapcss renderers, it works like this.


git-svn-id: https://josm.openstreetmap.de/svn/trunk@5214 0c6e7542-c601-0410-84e7-c038aed88b3b