Properly interpret more complex getCapabilities documents.
Keep a track of how deep into the XML tree reader has traversed. Do the
interpretation of the tags only if they are present at proper level.

If unknown tag is found, move the reader to the end of the tag, to avoid
confusion.

When using moveToReader within parseLayer, it is needed, to keep tagStack
updated.

This should fix problems with BE WMTS imagery - See: #12573



git-svn-id: https://josm.openstreetmap.de/svn/trunk@9887 0c6e7542-c601-0410-84e7-c038aed88b3b