Add script which adds replacements for svn:ignore and svn:externals
svn:ignore properties are added to the repository as .gitignore files (created by the `git svn create-ignore` command)
svn:externals properties are added as a git-svn clone of the referenced repository. The latest revision of the repository of the external that was created before the commit in the JOSM repository is used.