Retry tile download on transient errors.
Till now there was no strategy to retry tile download. Tiles for which download
finished with error where by design not downloaded again. Because we were
issuing surplus download requests it was not easy to notice, that some tiles
where downloaded at different zoom level. When surplus downloads where removed -
problem became aparent.

Redefine semantics in AbstractTileSourceLayer, that all Tiles that are not
loaded are tried again. Move all logic deciding whether tile has loaded or not
to TMSCachedTileLoaderJob.

Implment in TMSCachedTileLoaderJob following rules:
* when HTTP status code is between 400 and 500 - treat errors as permament
* when HTTP status code equals or is greater than 500 - then treat error as
temporary
* when there was exception in JCSCachedTileLoaderJob, then HTTP status code is
set to 599. Treat this as permament error. This is suboptimal, as we could treat
Socket read timeout and similar as temporary, and UnkownHostException as
permament

The other aproach is to implement retry count in Tile and retry for finite
number of times, preferably with some delay between tries.


Closes: #16747
See: #16743



git-svn-id: https://josm.openstreetmap.de/svn/trunk@14268 0c6e7542-c601-0410-84e7-c038aed88b3b